pyproject.toml
[project]
name = "crowd-evacuation"
version = "0.1.0"
description = "基于社会力模型和深度学习的人群疏散路径优化研究"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    # 核心计算
    "numpy>=1.24.0",
    "scipy>=1.10.0",

    # 可视化
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # 数据处理
    "pandas>=2.0.0",
    "tqdm>=4.65.0",

    # 配置
    "pyyaml>=6.0",

    # 机器学习
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",

    # 深度学习
    "torch>=2.0.0",

    # 强化学习
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "jupyterlab>=4.0.0",
]
